Added some error detection to simulation_parameters.py
Previously, using e.g. [target] instead of [[target]] in the simulation parameter file would have produced an error. Now, it is accepted and the input is automatically converted into a list with a single element, but the user is warned that they should change it since they'll still get an error if they try to define multiple targets using [target] rather than [[target]].